app/requirements.txt
alembic==1.12.1
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==1.1.2
flask==3.0.0
Flask-Migrate==4.0.5
flask-sqlalchemy==3.1.1
Flask-Validator==1.4.2
greenlet==3.0.1
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
isbnlib==3.10.3
iso3166==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.3.0
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
py-moneyed==0.8.0
pycountry==22.3.5
python-dotenv==1.0.0
pytz==2020.1
schwifty==2020.9.0
SQLAlchemy==2.0.23
typing-extensions==4.8.0
werkzeug==3.0.1
zipp==3.17.0
